consolidate functions to calculate rasterstats
