+++ pipelines/update_exposure.py
@@ -1,7 +1,14 @@
 from src.constants import ISO3S
-from src.datasources.floodscan import calculate_recent_flood_exposure_rasters
+from src.datasources import floodscan
 
 if __name__ == "__main__":
+
+    recent = False
+    clobber = False
+    verbose = False
+
     for iso3 in ISO3S:
         print(f"Processing {iso3}")
-        calculate_recent_flood_exposure_rasters(iso3)
+        floodscan.calculate_flood_exposure_rasters(
+            iso3=iso3, clobber=clobber, recent=recent, verbose=verbose
+        )
