+++ src/datasources/floodscan.py
@@ -3,11 +3,13 @@ from datetime import datetime
 from pathlib import Path
 from typing import Literal
 
+import pandas as pd
 import xarray as xr
+from sqlalchemy.engine import Engine
 from tqdm.auto import tqdm
 
-from src.datasources import worldpop
-from src.utils import blob
+from src.datasources import codab, worldpop
+from src.utils import blob, database
 
 DATA_DIR = Path(os.getenv("AA_DATA_DIR_NEW", ""))
 RAW_FS_HIST_S_PATH = (
@@ -22,12 +24,14 @@ RAW_FS_HIST_S_PATH = (
 )
 
 
-def calculate_recent_flood_exposure_rasters(
-    iso3: str, clobber: bool = False, verbose: bool = False
+def calculate_flood_exposure_rasters(
+    iso3: str,
+    clobber: bool = False,
+    recent: bool = True,
+    verbose: bool = False,
 ):
     """
-    Calculate recent flood exposure rasters for a given country.
-    Only looks to update data from the current year onwards.
+    Calculate flood exposure rasters for a given country.
 
     Parameters
     ----------
@@ -35,6 +39,8 @@ def calculate_recent_flood_exposure_rasters(
         ISO3 code of the country
     clobber: bool
         Whether to overwrite existing data
+    recent: bool
+        Whether to look only for data from the current year
     verbose: bool
         Whether to print progress of specific dates
 
@@ -51,11 +57,17 @@ def calculate_recent_flood_exposure_rasters(
         )
         if x.endswith(".tif")
     ]
+
     # filter to only this year onwards
-    this_year = datetime.today().year
-    recent_fs_raw_files = [
-        x for x in existing_fs_raw_files if f"300s_v{this_year}" in x
-    ]
+    if recent:
+        this_year = datetime.today().year
+        fs_raw_files = [
+            x for x in existing_fs_raw_files if f"300s_v{this_year}" in x
+        ]
+    # or check all files
+    else:
+        fs_raw_files = existing_fs_raw_files
+
     # check for existing processed exposure rasters
     existing_exposure_files = blob.list_container_blobs(
         name_starts_with=f"{blob.PROJECT_PREFIX}/processed/flood_exposure/"
@@ -64,7 +76,7 @@ def calculate_recent_flood_exposure_rasters(
 
     # stack up relevant raw Floodscan rasters
     das = []
-    for blob_name in tqdm(recent_fs_raw_files):
+    for blob_name in tqdm(fs_raw_files):
         date_in = datetime.strptime(
             blob_name.split("/")[-1][15:25], "%Y-%m-%d"
         )
@@ -118,6 +130,178 @@ def calculate_recent_flood_exposure_rasters(
         if verbose:
             print(f"uploading {blob_name}")
         blob.upload_cog_to_blob(blob_name, exposure.sel(date=date))
+
+
+def calculate_flood_exposure_rasterstats(
+    iso3: str,
+    engine: Engine,
+    clobber: bool = False,
+    verbose: bool = False,
+    output_table: str = "floodscan_exposure",
+):
+    """
+    Calculate flood exposure statistics from raster data for a given country.
+
+    This function processes flood exposure raster data for a specified country,
+    calculates exposure statistics at different administrative levels,
+    and stores the results in a PostgreSQL database.
+
+    Parameters
+    ----------
+    iso3 : str
+        Three-letter ISO country code
+    engine : sqlalchemy.engine.Engine
+        SQLAlchemy database engine for PostgreSQL connection
+    clobber : bool, optional
+        If True, reprocess existing dates. Default is False
+    verbose : bool, optional
+        If True, print additional processing information. Default is False
+    output_table : str, optional
+        Name of the output database table. Default is "floodscan_exposure"
+
+    Returns
+    -------
+    None
+        Results are written directly to the database
+
+    """
+    adm = codab.load_codab_from_blob(iso3, admin_level=2)
+    existing_exposure_rasters = [
+        x
+        for x in blob.list_container_blobs(
+            name_starts_with=f"{blob.PROJECT_PREFIX}/processed/"
+            f"flood_exposure/{iso3}/"
+        )
+        if x.endswith(".tif")
+    ]
+    existing_dates = database.get_existing_stats_dates(iso3, engine)
+    unprocessed_exposure_rasters = [
+        x
+        for x in existing_exposure_rasters
+        if datetime.strptime(x.split("/")[-1][13:23], "%Y-%m-%d")
+        not in existing_dates
+        or clobber
+    ]
+
+    # break list of exposure rasters into chunks, to avoid memory issues
+    # chunk length is arbitrary, but seems to work fine
+    chunk_len = 100
+    exposure_raster_chunks = [
+        unprocessed_exposure_rasters[x : x + chunk_len]
+        for x in range(0, len(unprocessed_exposure_rasters), chunk_len)
+    ]
+
+    # iterate over chunks
+    for exposure_raster_chunk in tqdm(exposure_raster_chunks):
+        # stack up exposure rasters in chunk
+        das = []
+        for blob_name in tqdm(exposure_raster_chunk):
+            date_in = datetime.strptime(
+                blob_name.split("/")[-1][13:23], "%Y-%m-%d"
+            )
+            try:
+                da_in = blob.open_blob_cog(blob_name)
+                da_in["date"] = date_in
+                da_in = da_in.persist()
+                das.append(da_in)
+            except Exception as e:
+                print(e)
+                print(f"couldn't open {blob_name}")
+
+        if len(das) == 0:
+            print("all complete for chunk")
+            continue
+        ds_exp_recent = xr.concat(das, dim="date").squeeze(
+            dim="band", drop=True
+        )
+        if verbose:
+            print(ds_exp_recent)
+
+        # iterate over admin level 2 regions and calculate exposure sums
+        dfs = []
+        for pcode, row in tqdm(
+            adm.set_index("ADM2_PCODE").iterrows(), total=len(adm)
+        ):
+            da_clip = ds_exp_recent.rio.clip([row.geometry])
+            dff = (
+                da_clip.sum(dim=["x", "y"])
+                .to_dataframe(name="total_exposed")["total_exposed"]
+                .astype(int)
+                .reset_index()
+            )
+            dff["ADM2_PCODE"] = pcode
+            dfs.append(dff)
+
+        df_exp_adm_new = pd.concat(dfs, ignore_index=True)
+        if verbose:
+            print(df_exp_adm_new)
+
+        # aggregate to admin levels and upload
+        df_exp_adm_new = df_exp_adm_new.merge(
+            adm[[x for x in adm.columns if "PCODE" in x]]
+        )
+        if verbose:
+            print("new raster stats calculated:")
+            print(df_exp_adm_new)
+
+        for adm_level in [0, 1, 2]:
+            if verbose:
+                print("aggregating to adm level:")
+                print(adm_level)
+            pcode_col = f"ADM{adm_level}_PCODE"
+            df_agg = (
+                df_exp_adm_new.groupby(["date", pcode_col])["total_exposed"]
+                .sum()
+                .reset_index()
+            )
+            df_agg["adm_level"] = adm_level
+            df_agg["iso3"] = iso3.upper()
+            df_agg = df_agg.rename(
+                columns={
+                    "total_exposed": "sum",
+                    pcode_col: "pcode",
+                    "date": "valid_date",
+                }
+            )
+            if verbose:
+                print("uploading to DB:")
+                print(df_agg)
+            df_agg.to_sql(
+                output_table,
+                schema="app",
+                con=engine,
+                if_exists="append",
+                chunksize=10000,
+                index=False,
+                method=database.postgres_upsert,
+            )
+
+
+def calculate_flood_exposure_rasterstats_regions(
+    region: dict,
+    engine: Engine,
+    output_table: str = "floodscan_exposure_regions",
+):
+    print(f"Processing {region['iso3']} region {region['region_number']}")
+    adm_stats_df = database.get_existing_adm_stats(region["pcodes"], engine)
+    region_stats_df = (
+        adm_stats_df.groupby("valid_date")["sum"].sum().reset_index()
+    )
+    region_stats_df["iso3"] = region["iso3"].upper()
+    region_stats_df["pcode"] = (
+        f'{region["iso3"]}_region_{region["region_number"]}'
+    )
+    region_stats_df["adm_level"] = "region"
+
+    region_stats_df.to_sql(
+        output_table,
+        schema="app",
+        con=engine,
+        if_exists="append",
+        chunksize=10000,
+        index=False,
+        method=database.postgres_upsert,
+    )
 
 
 def open_historical_floodscan():
