+++ src/utils/database.py
@@ -1,6 +1,7 @@
 import os
-from typing import Literal
+from typing import List, Literal
 
+import pandas as pd
 from dotenv import load_dotenv
 from sqlalchemy import (
     CHAR,
@@ -117,3 +118,26 @@ def postgres_upsert(table, conn, keys, data_iter, constraint=None):
     )
     conn.execute(upsert_statement)
     return
+
+
+def get_existing_stats_dates(iso3: str, engine) -> list:
+    query = f"""
+    SELECT DISTINCT valid_date
+    FROM app.floodscan_exposure
+    WHERE iso3 = '{iso3.upper()}'
+    ORDER BY valid_date
+    """
+    df_unique_dates = pd.read_sql(query, con=engine)
+    df_unique_dates["valid_date"] = pd.to_datetime(
+        df_unique_dates["valid_date"]
+    )
+    return df_unique_dates["valid_date"].to_list()
+
+
+def get_existing_adm_stats(pcodes: List[str], engine) -> pd.DataFrame:
+    query = f"""
+    SELECT *
+    FROM app.floodscan_exposure
+    WHERE pcode IN ({", ".join([f"'{pcode}'" for pcode in pcodes])})
+    """
+    return pd.read_sql(query, con=engine)
